# Conflicts:
#	Projet/PTUT/Neuron.cpp

+++ Projet/PTUT/Neuron.cpp
@@ -11,7 +11,7 @@ namespace SOM
 		m_nWeightDim = weightDim;
 		m_fWeight = new uchar[m_nWeightDim];
 		for (uint i = 0; i < m_nWeightDim; ++i) {
-			m_fWeight[i] = (uint)(rand() % 255);
+			m_fWeight[i] = (uchar)(rand() % 255);
 			printf("%d", m_fWeight[i]);
 		}
 		/*uchar weight = this->GetWeight(0);
